% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{aPage}
\alias{aPage}
\title{Page}
\usage{
aPage(..., title = aLogo(), window_title = "Astho")
}
\arguments{
\item{...}{Content of the page, passed to \code{\link[shiny:navbarPage]{shiny::navbarPage()}}.}

\item{title}{Title of the application.}

\item{window_title}{Title of the application on browser tab.}
}
\description{
Astho themed page.
}
