% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{asAgencyLookup}
\alias{asAgencyLookup}
\title{Agency Lookup}
\usage{
asAgencyLookup(button)
}
\arguments{
\item{button}{An \code{\link[=actionButton]{actionButton()}}.}
}
\description{
Make agency lookup button from an existing \code{\link[=actionButton]{actionButton()}}.
}
