% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{asthoActionButton}
\alias{asthoActionButton}
\title{Astho Button}
\usage{
asthoActionButton(..., class = "")
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:actionButton]{shiny::actionButton()}}.}

\item{class}{Additional class(es) to pass to the button.}
}
\description{
Astho-themed button.
}
