% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{asthoApp}
\alias{asthoApp}
\title{Atho themed application}
\usage{
asthoApp(
  ...,
  header = NULL,
  title = aLogo(),
  window_title = "Astho",
  collapsible = TRUE
)
}
\arguments{
\item{...}{Content of the page, passed to \code{\link[shiny:navbarPage]{shiny::navbarPage()}}.}

\item{header}{Additional headers.}

\item{title}{Title of the application.}

\item{window_title}{Title of the application on browser tab.}

\item{collapsible}{Whether to collapse the navbar on small screens.}
}
\description{
Astho themed application.
}
