% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{aTheme}
\alias{aTheme}
\title{Theme}
\usage{
aTheme(
  ...,
  primary = "#06476b",
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL
)
}
\arguments{
\item{...}{Passed to \code{\link[bslib:bs_theme]{bslib::bs_theme()}}.}

\item{primary}{Primary color passed to \code{\link[bslib:bs_theme]{bslib::bs_theme()}}.}
}
\description{
Default Astho theme.
}
