% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{coordinates_to_state}
\alias{coordinates_to_state}
\title{Coordinates to State}
\usage{
coordinates_to_state(coordinates)
}
\arguments{
\item{coordinates}{Value of \code{input$coordinates}.}
}
\value{
\code{NA} if the user is outside the United States.
}
\description{
Transforms coordinates as set by \code{input$coordinates} to a state code.
}
\keyword{internal}
