% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo_locate}
\alias{geo_locate}
\title{Geolocate}
\usage{
geo_locate(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A shiny session.}
}
\description{
Uses the navigator's geolocation API to get the coordinates of the user.
This is to be used within shiny, the coordiantes are set on
\code{input$coordinates}.
}
