% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_state}
\alias{get_state}
\title{Get State}
\usage{
get_state(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A shiny session.}
}
\description{
Get state, result of the \code{\link[=geo_locate]{geo_locate()}} prompt.
}
