% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{hide_footer}
\alias{hide_footer}
\title{Hide Footer}
\usage{
hide_footer(id, tabs, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of footer element.}

\item{tabs}{List of tab names or their \code{value} (if defined) of
tabs on which to hide the footer.}

\item{session}{A valid shiny session.}
}
\description{
Hide a footer in specific tabs.
}
