% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{inject_footer}
\alias{inject_footer}
\title{Inject Footer}
\usage{
inject_footer(footer, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{footer}{Footer to inject.}

\item{session}{A valid shiny session.}
}
\description{
Hide a footer in specific tabs.
}
