% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{shortcut}
\alias{shortcut}
\title{Shortcut}
\usage{
shortcut(button, tab, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{button}{ID of button to trigger}

\item{tab}{Name of tab to trigger.}

\item{session}{Shiny session.}
}
\description{
Triggers a button click on another tab.
}
\examples{
library(shiny)

ui <- asthoApp(
  tabPanel(
    "HOME",
    aPage(
      left = div(
        h1("Sidebar")
      ),
      center = div(
        h1("Main page"),
        actionButton("shortcut", "Shortcut")
      ),
      right = div(
        storiesOutput("stories")
      )
    )
  ),
  tabPanel(
    "ACTIVITIES",
    actionButton("story", "Story")
  )
)

server <- function(input, output, session) {
  observeEvent(input$story, {
    print("story")
  })

  observeEvent(input$shortcut, {
    shortcut("story", "ACTIVITIES")
  })
}

if(FALSE)
  shinyApp(ui, server)

}
