% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{stories}
\alias{stories}
\title{Stories}
\usage{
stories(...)
}
\arguments{
\item{html}{HTML to display (\link[htmltools:builder]{htmltools::tags}).}
}
\description{
Render stories as stacked cards.
}
\examples{
library(shiny)

ui <- fluidPage(
 storiesOutput("id")
)

server <- function(input, output){
 output$id <- renderStories({
   stories(h2("Hello"))
 })
}

if(interactive())
 shinyApp(ui, server)

}
