% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{treeInputSearch}
\alias{treeInputSearch}
\title{TreeInput with search}
\usage{
treeInputSearch(
  inputId,
  ...,
  placeholder = "search",
  button = shiny::icon("search"),
  button_class = ""
)
}
\arguments{
\item{inputId}{passed to \code{shinyWidgets::treeInput()}.}

\item{...}{passed to \code{shinyWidgets::treeInput()}.}

\item{placeholder}{Placeholder text.}

\item{button}{Content of the button.}

\item{button_class}{Additional classes to pass to the button.}
}
\description{
\code{shinyWidgets::treeInput()} with search.
}
