% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{aTheme}
\alias{aTheme}
\alias{aDeps}
\alias{serve_assets}
\title{Theme}
\usage{
aTheme(
  ...,
  primary = "#06476b",
  secondary = "#DD5D2A",
  success = "#6BBE89",
  info = NULL,
  warning = NULL,
  danger = NULL
)

aDeps()

serve_assets()
}
\arguments{
\item{...}{Passed to \code{\link[bslib:bs_theme]{bslib::bs_theme()}}.}

\item{primary, secondary, success, info, warning, danger}{Colors.}
}
\description{
Default Astho theme.
}
